pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vidforge"
version = "0.1.0"
description = "Modular video generation system with multiple sources and platform targets"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "sf-hamilton[visualization]>=0.30",
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "pillow>=10.0",
    "rembg[cpu]>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "httpx>=0.25",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
all = [
    "vidforge[youtube]",
    "vidforge[dev]",
    "vidforge[music]",
]
youtube = [
    "google-api-python-client>=2.100",
    "google-auth-httplib2>=0.1",
    "google-auth-oauthlib>=1.0",
]
music = [
    "yt-dlp>=2023.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.15",
    "pre-commit>=4.0",
    "mypy>=1.8",
]

[project.scripts]
vidforge = "vidforge.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "TID",  # flake8-tidy-imports (single import per line)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # typer.Argument/Option in defaults (standard Typer pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["vidforge"]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.flake8-import-conventions]
banned-from = ["*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers -x"

[tool.coverage.run]
source = ["src/vidforge"]
branch = true
omit = ["*/cli.py", "*/__init__.py", "*/assets/bg_remove.py", "*/pipeline.py"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "rembg.*",
    "sf_hamilton.*",
    "hamilton.*",
    "yaml.*",
    "scipy.*",
    "yt_dlp.*",
    "google.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["vidforge.pipeline"]
ignore_errors = true

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    '"""TODO"""',
]
